+++ prisma/migrations/20250708160113_added_password_protection_for_links/migration.sql
@@ -1,0 +1,6 @@
+-- AlterTable
+ALTER TABLE "ResetPasswordToken" ALTER COLUMN "expDate" SET DEFAULT NOW() + INTERVAL '15 minutes';
+
+-- AlterTable
+ALTER TABLE "ShortUrl" ADD COLUMN     "password" TEXT,
+ALTER COLUMN "expDate" SET DEFAULT NOW() + INTERVAL '5 days';