+++ src/app/api/cron/delete-records/route.ts
@@ -6,46 +6,49 @@ export async function GET() {
     const expired30Days = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000); // 30 dias em milissegundos
     console.log(`Cron job ran at ${now.toISOString()}`);
     const errors: { message: string, reason: string }[] = [];
-    
+
     // Delete expired data analytics
     let deletedDataAnalytics = 0;
-    try {
-        const deleted = await prisma.dataAnalytics.deleteMany({
-            where: {
-                accessedAt: {
-                    lt: expired30Days,
-                },
-            },
-        });
-
-        deletedDataAnalytics = deleted.count
-
-    } catch (error) {
-        console.error("Error deleting expired data analytics:", error);
-        errors.push({ message: "Error deleting expired data analytics", reason: JSON.stringify(error) || "Could not set reason, check console" });
-    }
-
-    // Delete expired links
     let deletedLinks = 0;
     try {
-        const deleted = await prisma.shortUrl.deleteMany({
-            where: {
-                expDate: {
-                    lt: now,
+        await prisma.$transaction(async (tx) => {
+            const deletedDA = await tx.dataAnalytics.deleteMany({
+                where: {
+                    accessedAt: {
+                        lt: expired30Days,
+                    },
                 },
-            },
+            });
+
+            const deletedPublicSU = await tx.shortUrl.deleteMany({
+                where: {
+                    expDate: {
+                        lt: now,
+                    },
+                    userId: null,
+                },
+            });
+
+            const deletedPrivateSU = await tx.shortUrl.deleteMany({
+                where: {
+                    expDate: {
+                        lt: now,
+                    },
+                    userId: {
+                        not: null,
+                    },
+                    dataAnalytics: {
+                        none: {},
+                    },
+                },
+            });
+        
+            deletedLinks = deletedPublicSU.count + deletedPrivateSU.count;
+            deletedDataAnalytics = deletedDA.count
         });
-
-        console.log(`Deleted ${deleted.count} expired links`);
-        deletedLinks = deleted.count;
-
     } catch (error) {
-        console.error("Error deleting expired links:", error);
-        errors.push({ message: "Error deleting expired links", reason: JSON.stringify(error) || "Could not set reason, check console" });
-    }
-
-    if(errors.length > 0) {
-        return NextResponse.json({ errors: errors }, { status: 500 });
+        console.error("Error deleting expired records:", error);
+        errors.push({ message: "Error deleting records", reason: JSON.stringify(error) || "Could not set reason, check console" });
     }
 
     return NextResponse.json({ deleted: true, links: deletedLinks, data_analytics: deletedDataAnalytics }, { status: 200 });